Fix creation of XPC interface
For the XPC interface to work, the static registrar must
be used, the protocol must have an explicit name, that name
(as a string) must be used to retrieve the protocol, and the
protocol's interface type must be implemented by a class
(not just referenced) in the main app binary. Ugh.